Improved algorithm for bs4 search and added function to fetch makers.
All possible car makers and models are fetched from finn and used to create
the search URL, so it is easier to input car searches. Also fuel types and
karosseri are available.

Use of bs4 to find cars and data is improved.

Experimenting with fetching detailed car data. Not really sure what to get
here, of maybe get all?